requirements.txt
Pillow==9.1.0
requests==2.27.1
urllib3==1.26.9
